Serve landing page via templates
