+++ backend/app/static/css/styles.css
@@ -1,0 +1,271 @@
+:root {
+  color-scheme: light dark;
+  --bg: radial-gradient(circle at top, #f3f5ff 0%, #eef3ff 35%, #ffffff 100%);
+  --card-bg: rgba(255, 255, 255, 0.85);
+  --border: rgba(120, 136, 189, 0.35);
+  --text-main: #1f273d;
+  --text-muted: #4a5674;
+  --accent: #4450d6;
+}
+
+body {
+  margin: 0;
+  font-family: "Inter", "Segoe UI", system-ui, sans-serif;
+  background: var(--bg);
+  color: var(--text-main);
+  min-height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 3rem 1.5rem 4rem;
+}
+
+main {
+  width: min(860px, 100%);
+  background: var(--card-bg);
+  border: 1px solid var(--border);
+  border-radius: 18px;
+  box-shadow: 0 20px 45px rgba(80, 98, 160, 0.18);
+  backdrop-filter: blur(14px);
+  padding: 2.5rem clamp(1.5rem, 3vw, 3rem);
+  line-height: 1.7;
+}
+
+h1 {
+  font-size: clamp(2.2rem, 4vw, 2.8rem);
+  margin: 0 0 0.2rem;
+}
+
+h2 {
+  font-size: clamp(1.2rem, 3vw, 1.5rem);
+  margin-top: 2.2rem;
+  margin-bottom: 0.6rem;
+}
+
+p.lead {
+  margin: 0;
+  color: var(--text-muted);
+  font-size: 1.05rem;
+}
+
+section {
+  border-top: 1px solid var(--border);
+  padding-top: 1.6rem;
+}
+
+ul,
+ol {
+  padding-left: 1.1rem;
+  margin: 0.6rem 0;
+}
+
+li {
+  margin-bottom: 0.35rem;
+}
+
+code {
+  background: rgba(68, 80, 214, 0.08);
+  color: var(--accent);
+  padding: 0.15rem 0.4rem;
+  border-radius: 6px;
+  font-size: 0.95rem;
+}
+
+a {
+  color: var(--accent);
+  font-weight: 600;
+  text-decoration: none;
+}
+
+a:hover {
+  text-decoration: underline;
+}
+
+.cta-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(240px, 1fr));
+  gap: 1.2rem;
+  margin-top: 1.4rem;
+}
+
+.chat-section {
+  margin-top: 2rem;
+}
+
+.chat-card {
+  border: 1px solid var(--border);
+  border-radius: 16px;
+  padding: 1.2rem clamp(1rem, 2vw, 1.6rem);
+  background: rgba(255, 255, 255, 0.7);
+  box-shadow: 0 10px 30px rgba(69, 86, 150, 0.12);
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.chat-log {
+  max-height: 320px;
+  overflow-y: auto;
+  padding-right: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.msg {
+  padding: 0.6rem 0.8rem;
+  border-radius: 12px;
+  line-height: 1.55;
+}
+
+.msg strong {
+  margin-right: 0.3rem;
+}
+
+.msg-text {
+  white-space: pre-wrap;
+}
+
+.msg-user {
+  align-self: flex-end;
+  background: rgba(68, 80, 214, 0.12);
+}
+
+.msg-assistant {
+  background: rgba(255, 255, 255, 0.85);
+  border: 1px solid rgba(68, 80, 214, 0.1);
+}
+
+.msg-focus {
+  margin-top: 0.4rem;
+  font-size: 0.9rem;
+  color: var(--text-muted);
+  display: inline-flex;
+  flex-wrap: wrap;
+  gap: 0.35rem;
+  align-items: baseline;
+  background: rgba(68, 80, 214, 0.08);
+  padding: 0.35rem 0.6rem;
+  border-radius: 8px;
+}
+
+.msg-spoken {
+  margin-top: 0.4rem;
+  font-size: 0.9rem;
+  color: var(--text-muted);
+  display: flex;
+  align-items: center;
+  gap: 0.35rem;
+}
+
+.msg-spoken audio {
+  margin-left: 0.35rem;
+  height: 32px;
+}
+
+.msg-feedback {
+  margin-top: 0.4rem;
+  font-size: 0.9rem;
+  color: var(--text-muted);
+}
+
+.msg-feedback ul {
+  margin: 0.3rem 0 0;
+  padding-left: 1.1rem;
+}
+
+.chat-form textarea {
+  width: 100%;
+  border: 1px solid var(--border);
+  border-radius: 12px;
+  padding: 0.75rem;
+  font-size: 1rem;
+  resize: vertical;
+  min-height: 90px;
+  font-family: inherit;
+}
+
+.chat-actions {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 0.6rem;
+  gap: 0.75rem;
+  flex-wrap: wrap;
+}
+
+.chat-buttons {
+  display: flex;
+  gap: 0.5rem;
+  flex-wrap: wrap;
+}
+
+.chat-actions button {
+  background: var(--accent);
+  color: #fff;
+  border: none;
+  padding: 0.6rem 1.2rem;
+  border-radius: 10px;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: transform 0.1s ease;
+}
+
+.chat-actions button.secondary {
+  background: rgba(68, 80, 214, 0.12);
+  color: var(--accent);
+}
+
+.chat-actions button:disabled {
+  opacity: 0.6;
+  cursor: progress;
+}
+
+.chat-actions button:not(:disabled):hover {
+  transform: translateY(-1px);
+}
+
+.chat-status {
+  font-size: 0.9rem;
+  color: var(--text-muted);
+}
+
+.lao {
+  font-weight: 600;
+  font-size: 1.05rem;
+}
+
+.sr-only {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  padding: 0;
+  margin: -1px;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  border: 0;
+}
+
+.card {
+  padding: 1.2rem 1.4rem;
+  border-radius: 12px;
+  border: 1px solid var(--border);
+  background: rgba(255, 255, 255, 0.65);
+  box-shadow: 0 10px 25px rgba(94, 113, 178, 0.08);
+}
+
+footer {
+  margin-top: 2.6rem;
+  font-size: 0.92rem;
+  color: var(--text-muted);
+}
+
+@media (max-width: 640px) {
+  body {
+    padding: 2rem 1rem 3rem;
+  }
+
+  main {
+    padding: 2rem 1.2rem;
+  }
+}
