+++ backend/requirements.txt
@@ -2,5 +2,6 @@ fastapi>=0.119,<1.0
 uvicorn[standard]>=0.27,<1.0
 pydantic>=1.10,<2.0
 numpy>=1.24,<2.0
+jinja2>=3.1,<4.0
 python-multipart>=0.0.9
 httpx>=0.27,<1.0
